requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
# Copyright (c) 2020 Spirent Communications
#
# All rights reserved. This program and the accompanying materials
# are made available under the terms of the Apache License, Version 2.0
# which accompanies this distribution, and is available at
# http://www.apache.org/licenses/LICENSE-2.0

PyYAML # MIT
urllib3 # MIT
GitPython<2.1.12;python_version<'3.0' # BSD License (3 clause)
GitPython;python_version>='3.0' # BSD License (3 clause)
requests!=2.20.0 # Apache-2.0
netaddr # BSD
openpyxl
